util: play: use SoundFounts 2 !! compatibility !!
The old way of playing is no more supported.
This may raise compatibility problems when the soundfont URL has been
memorized in a cookie by the editor.
In this case, delete the cookie or force the URL to 'Scc1t2'.

FossilOrigin-Name: 9fad90c959849bacf797eda5fde4a2acb2c7efae